works sometimes, lots of leaks

+++ Star.h
@@ -5,33 +5,33 @@
 #include "List.h"
 
 class Starlist{
-	private: 
-		int size;	
+	private:
+		unsigned int size;
 	public:
 		List * list;
 		Starlist();
 		~Starlist();
 		long addPlanet();
-		bool removePlanet(int);
-		Planet * getPlanet(int);
+		bool removePlanet(long);
+		Planet * getPlanet(long);
 		void orbit();
 		void printStarInfo();
-		unsigned getCurrentNumPlanets();	
+		unsigned getCurrentNumPlanets();
 };
 
 class Starvector{
 	private:
-		int current_planets;	
+		int current_planets;
 	public:
 		Vector * vector;
 		Starvector();
 		~Starvector();
 		long addPlanet();
-		bool removePlanet(int);
-		Planet * getPlanet(int);
+		bool removePlanet(long);
+		Planet * getPlanet(long);
 		void orbit();
 		void printStarInfo();
-		unsigned int getCurrentNumPlanets();	
+		unsigned int getCurrentNumPlanets();
 };
 
-#endif 
+#endif
